Patch from Lars E. Pettersson, sm6rpz
The code should work on IC-751 and IC-751A, both with the original Icom
UX-14 board (an option needed to be able to use CI-V) and the Piexx
UX-14px-board.

The Piexx UX-14px board have the same capabilities as the Icom one, but
adds S-meter reading, and PTT (can not read if the PTT is set though.)

The Icom IC-751 had FM as an option, the IC-751A had FM built in.

I had to add an extra if-statement in frame.c for the CW-narrow to work.
The bandwidth numbers used by Icom seem to be a jungle...


git-svn-id: https://hamlib.svn.sourceforge.net/svnroot/hamlib/trunk@2070 7ae35d74-ebe9-4afe-98af-79ac388436b8