K3 get extra level for IF center frequency
Implemented rig_get_ext_level() for the K3 FI command which returns the
last four digits of the IF center frequency.  This value changes as the
operator adjusts the shift control, changes mode, or adjusts the bandwidth
control.  The function returns the actual frequency of 8210000 Hz plus
the value returned by the FI command.

To get this frequency value the token IFCTR must be passed to
rig_get_ext_level().  This token is specific to the K3 backend.

git-svn-id: https://hamlib.svn.sourceforge.net/svnroot/hamlib/trunk@3029 7ae35d74-ebe9-4afe-98af-79ac388436b8