Add support for TX Query on K2 and K3
The K2 and K3 support a TQ command to query the transmit status.  As
only four bytes are returned it is much faster than the longer IF
command response.  Access is via rig_get_ext_level--rigctl 'l' command
with a Level mnemonic of 'txst'.  Returns '0' on RX and '1' on TX.

git-svn-id: https://hamlib.svn.sourceforge.net/svnroot/hamlib/trunk@3032 7ae35d74-ebe9-4afe-98af-79ac388436b8