Added attenuator, preamplifier and RSSI to FUNcube
The Attenuator/Preamp has been added: not all the LNA values have been
added because Hamlib can only have maximum 8 values and FUNcube has
more.

The RSSI readout is performed on the IF and it is not calibrated. To
calibrate it, it would be necessary to read all the gains in the
complete chain to compute the real RSSI.

The driver now also supports firmware versions 1.x.

Author: Stefano Speretta, Innovative Solutions In Space BV
<s.speretta@isispace.nl>

Signed-off-by: Stefano Speretta <s.speretta@isispace.nl>

git-svn-id: https://hamlib.svn.sourceforge.net/svnroot/hamlib/trunk@3053 7ae35d74-ebe9-4afe-98af-79ac388436b8