Support high resolution S-meter on K3
Recent firmware versions since 4.37 now offer support for higher
resolution for the S-meter output via the SMH command.  The K3 backend
performs the STRENGTH level routine and passes other levels to the
Kenwood backend.

This patch has not been tested with a firmware version older than 4.39
to assure that operation with the SM command still works as expected.