Fix error for 'make check'
Implement several simple shell scripts to exercise various aspects of
Hamlib.  The previous implementation relied on the installation of
DejaGNU and the tests/config and tests/*.test directories were not
distributed.  Also, later Autotools documenation seems to suggest that
DejaGNU is to be replaced by Autotest.  The generation and use of simple
scripts was chosen to provide a usable check target until Autotest can
be supported.