Provide libltdl in source tarballs
Initializing libltdl is now done by the LTDL_INIT macro in configure.ac.
This now creates a libltdl directory in the tree (not tracked in Git)
when autoreconf is run.  The downside is that this complicates the build
a little bit with the advantage that the build can now take place on a
system that does not have libltdl-dev installed as Hamlib will build on
such systems by passing the '--with-included-ltdl' flag to the configure
script.  In particular, this simplifies the build on Linux for the mingw32
host by not having to run libtoolize manually and assures that libltdl
is built with the same build and host architectures as Hamlib.

Now check if '--with-included-ltdl' has been passed to configure and if
not, only then test for the presence of an installed ltdl.h.

Commented out AM_MAINTAINER_MODE macro.  It is recommended by the
Automake manual that this macro and its configure options not be used as
the Autotools system is designed to regenerate its files.  Let me know
if this is problem as it can be reversed easily enough.  The autogen.sh
script no longer passes the '--enable-maintainer-mode' option to
configure when bootstrapping the build system from Git.

Minor edits in configure.ac to improve readability.

Updated build-win32 files in scripts/