ABI backward compatibility enhancement
Use the libtool link mode argument '-version-info' to set the Hamlib 3
ABI to be compatible with Hamlib 1.2 ABI.  As the current ABI has not
been changed or removed, we should be able to maintain backward
compatibility.  The libtool docs state that so long as the ABI is only
added to, backward compatibility should be maintained.

My idea here is to be able to have presently compiled programs be able
to use binary libraries of Hamlib 3 without recompiling.  So far I have
tested this locally on Linux and Windows.  More testing is welcome.