configure.ac: rigmem XML support user optional
I found that building on Cygwin that when libxml2 is installed, MSYS
fails the compile as libxml2 is installed outside the MSYS idea of
'root'.  As this feature depends on an external library, give user the
option of compiling with this support.  Default is to not build rigmem
with XML support.