Install build files rather than symlink
Previously, invoking autogen.sh would create symbolic links to needed
Autotools files in the tree only when working from a Git checkout
(source distributions have the files included).  Consideration of this
included the possibility of bugs if the Autotools packages were upgraded
on the developer's system that strange bugs could result.  Copying the
files means the tree is self contained until the next time autoreconf or
autogen.sh is run.

The top level Makefile.am included the variable $(subdirs) in both
SUBDIRS and DIST_SUBDIRS assignments.  While this variable may contain a
useful value, it is undocumented in the manuals.  Relying on it may
result in some directory being ommitted or other changes in the future,
so removing it from the assignments.