rigmatrix.html target is conditional in tests/Makefile
If libgd-dev is not installed, an Automake conditional will disable the
rigmatrix.html target in tests/Makefile.  Optionally, this target may
also be disabled by use of the '--disable-html-matrix' option passed to
the configure script.  By default the configure script will enable the
target if the libgd development files are found and disable it when they
are not.