IC-746/756, IC-PCR8500 fixes
* The 746 has an APF, wasn't in its configuration, added, tested, works.

* The 756Pro has an ANF function, wasn't in its configuration, added,
tested, works.

* The 756Pro's signal strength conversion table was way wrong.
I noticed this while making scope plots in my program JRX --
the strong signals using the STRENGTH output would all flatten
out at a rather low level, but the RAWSTR output showed the
full range. Changed the conversion table, tested, fixed.

R8500 notes:

The R8500 on/off functions work differently than the other Icoms,
in fact it's not too strong to say that this radio is a breed apart.
Most Icom radios have a group command code for on/off functions (16),
then a code for the function, then a flag for on or off. The R8500
combines the second and third into one, and this required a bit of
recoding in "icom.c". Interestingly, someone had created some special
designators in "icom_defs.h" for the R8500, but then never coded them
to the degree that they worked. Now they work. :)

So the AGC, NB and APF functions all work now.

The R8500 doesn't have a preamp control, so I took that out.

The R8500 does have IF shift and a 3-step attenuater, so I put those in.

Signed-off-by: Nate Bargmann <n0nb@n0nb.us>