WinRadio G313: Updates from Julian Campbell
"I've re-written the hamlib backend to stream the audio, if and spectrum
pcm samples to simple file system objects.

If no path is given, streaming is not requested. If the path can't be
openned, no streaming is requested.

The file is openned non-blocking and overruns are just silently
discarded.

I've tested it using named fifos and if there is no active reader it
fails to open and if there is an active reader it streams fine.

Thats as simple as it can get.

In terms of the project files, I've put the wrg313api.c and wrg313api.h
files in the existing linradio subfolder of winradio.

The .h file is the g313 equivalent of the wrapi.h file already there.
The .c file actually just contains the dlopen and the assignments to the
function variables.

The actual API shared library is binary only (and 32-bit only), and not
included in my patch."

Signed-off-by: Nate Bargmann <n0nb@n0nb.us>