pcr.c, pcr1500.c: Add DSP support
"[This patch] allows PCR radios to use DSP units correctly for the first
time under Hamlib.

We're all familiar with a typical DSP presentation on a radio -- you can
turn it on and off, you can adjust the degree of noise processing, and
you can sometimes even enable a notch filter. In the PCR radios
that have a DSP installed, none of this worked until now.

I corrected the code in pcr.c, changes that apply to all the PCR radios,
and in pcr1500.c, the radio I have, but someone else will have to
edit the files for the radios I don't own so the user interface works
right. I decided it would be unwise to try to edit files for radios
I couldn't test.

I just re-read the above, and I need to add that the changes to prc.c
shouldn't adversely affect the PCR models other than the one I own."

Signed-off-by: Nate Bargmann <n0nb@n0nb.us>