gs232b.c: Improve get_position string parsing
I've found a problem with my setup. I'm using a Yaesu G-5500 with an
arduino running:

http://blog.radioartisan.com/yaesu-rotator-computer-serial-interface/

to emulate a GS232b.

The problem is that the Elevation in rotctl was incorrectly reported,
with the least significant digit show. The problem I've found is some
incompatibility between the format of the answer and the parser. Here
I'm pasting a new version of gs232b_rot_get_position() which both solves
my issue and should be a bit more reliable to protocol differences,
while simpler than the previous.

(a white space in sscanf() matches none or any number of whitespaces, my
emulated GS232B was reporting AZ=123EL=033 with no spaces. This got
broken recently, and I haven't touched the code running in the arduino).

Signed-off-by: Nate Bargmann <n0nb@n0nb.us>