Fix Yaesu newcat backend set_tx_vfo issue.
The newcat Yaesu backend was using a toggle command to set the TX VFO
where an explicit set command is avaiable.

This fix corrects the function for the FT-2000(D) and the FT-DX5000.

There is still an issue with the FT-450D which as far as I can see has
no definitive way of setting which VFO will be used for TX.

Signed-off-by: Nate Bargmann <n0nb@n0nb.us>