Fix Kenwood TS-870s split and bandwidth set/get.
The TS-870s has no width set for SSB and AM, instead the bandwidth is
set via a high and low pass fiter setting. The backend now sets these
to try and achieve the requested bandwidth. The HPF setting is assumed
to be the default.

The bandwidth query reads both the HPF and the LPF and returns the
difference.

The value of RIG_BANDWIDTH_NORMAL skips bandwidth adjustements.

The get and set_split operations have been enabled.