Get firmware revision for TS590s in rig_open.
Defects in the TS590s f/w can be addressed selectively according to
revision.

Enable kenwood_open function for TS590s.

For some reason the TS590s did not use the kenwood_open function, as
fetching the firmware revision may be required for this rig and it is
done in that function; it has been enabled.