Fix a buffer overflow in the Windows termios emulation.
The  YACK() macro  generates  a  message of  variable  length due  the
__FILE__ string  and a  message returned  by MS  FormatMessage() being
inserted  into a  fixed length  string  buffer. I  have increased  the
buffer length to 1024 bytes and used snprintf to avoid overflows.

It looks like the whole YACK() macro should do nothing unless DEBUG is
defined so  there is probably  a small performance  gain to be  had by
doing that also. I have left that to the experts to decide.