Implement rig_{set,get}_split_freq_mode() for IC-756 series.
This implementation reduces  the number of VFO excahnges or  sets to a
minimum when setting the TX VFO frequency and mode together.

This also  includes a  much smarter implementation  of the  logic used
when the  new config option 'no_xchg'  is enabled.  Now it  is able to
accurately  determine  the  original   VFO  selected  and  leaves  the
sequences with the  same VFO selected.  There is one  corner case when
both VFOs are identical (frequency, mode and width) at the start where
it has  to take a  guess as to the  selected originally VFO  (Icom CAT
protocol has no  command to elicit this information). In  that case it
guesses as the last VFO  selected using Hamlib.  Applications can gain
absolute certainty  by setting the  VFO themselves prior to  using the
rig_{set,get}_{split,}_{freq,mode,freq_mode}() function  series.  This
now gives seamless  TX VFO configuration without  any relay clattering
or TX/RX audio glitches.