Fix location of custom Python bindings.
I have a small patch to correct the install location of the python
bindings.

Since it's not a pure python extension is belongs in pyexecdir instead
of pythondir, which for multilib systems like Fedora will get installed
into /usr/lib64/python{ver}/site-packages on 64bit systems.

This is consistent with the automake documentation:

http://www.gnu.org/software/automake/manual/html_node/Python.html

(Patch adjusted by n0nb for Hamlib 3.0 bindings/Makefile.am)

Signed-off-by: Nate Bargmann <n0nb@n0nb.us>