Add support for DMK Engineering URI
I have a problem with the CM119A GPIO. The PTT line of the DMK
Engineering URI was not activated by the hamlib  when using fldigi or
rigctl.

I fiddled with the source code and managed to edit the code in cm108.c
to get the PTT-line activate on my URI-box.
A successful rigctl command is:

rigctl -p /dev/hidraw0 -C ptt_pathname=/dev/hidraw0,ptt_type=CM108,ptt_bitnum=2
Attached is the modified source code file and a patch in 'git diff'
format.

The version of origin is 1.2.15.3.

73 de Veijo OH3NFC

Signed-off-by: Nate Bargmann <n0nb@n0nb.us>