Properly exclude non-Linux/MS Windows build of G313
While the Winradio G313 backend wasn't being compiled on a non-Linux
POSIX system, the register function in winradio.c was which caused an
error linking rigctl.  Also, the host_os test in the configure script
now looks for a string containg "*linux-gnu*" which accepts such systems
as the Raspberry Pi which is defined as "linux-gnueabihf".

Thanks to Lorenzo Simoncello, IW3HER, for reporting this build error on
the Raspberry Pi.