Yaesu newcat style command response reading improvements
After testing on an FT-450D which tends to send busy responses "?;" to
subsequent commands after a band  changing command it appears that the
undoumented "?;"  rig busy response  should be  treated as a  wait and
subsequent reads may return the  expected response i.e. the command is
still in progress and should not be resent.

As this is  all based on observations of failure  modes rater than any
doumentation from Yaesu, there may be further enhancements necessary.