Increase communications timeout for Icom rigs
Icom  users, who  use  the  K9JM CI-V  router  to  protect their  PW-1
amplifiers from  damage due to  Icom's broken CI-V  protocol handling,
require a 1s delay  after opening the serial port. This  is due to the
router  using  an  Arduino  &  Arduino  USB  adapter  that  employs  a
bootloader which  holds the RS-232  interface for 1s after  reset (DTR
releases reset).   Hamlib already  has a  retry and  timeout mechanism
which was  set to 3 and  200ms respectively for most  Icom rigs.  This
change increses that  timeout to 1s. The increaased  timeout will have
little or no impact on clients as  it only comes into play if there is
a problem.