Kenwood and Elecraft rigs sometimes ignore commands
These  rigs  can  opt  to  ignore set  type  commands  when  they  are
busy. This means  that sending a set command like  FAnnnnnnnnn; to set
the frequency can  fail. The rig should send "?;"  when busy but since
no reply is normally expected it is not convenient to read for a reply
after a  set command since it  will block until timeout.   This change
sends an "AI;" command after a  command that is not expected to reply,
this should guarantee a reply of  some sort allowing any busy reply to
be read  without blocking  and initiate  a retry  of the  original set
command. The  "AI;" command is chosen  because it is available  on all
rigs AFAIK and at least on Elecraft is guaranteed to be processed even
when the rig is busy.