Removed kenwood_cmd function and replaced uses of it
The function kenwood_cmd doesn't add value  and it uses a reply buffer
that is  not required.  Also  the use of  a reply buffer  disables the
verification  of commands  that  have no  reply feature.   Substituted
kenwood_simple_cmd calls  that do  the job  correctly without  a reply
buffer.

Added delay to busy retry loop when sending commands to rig.

Implemented retries on wrong length repies from rig also with a delay.