Allow for failed test command on the K2
The  K2 initialization  was checking  if  RTTY mode  was available  by
sending an  MD6; command  and they  querying the mode  to check  if it
worked. The latest kenwood routines now  fail if an invalid command is
sent to the K2 initialization need only try the MD6; command to see if
RTTY is available on the rig.