Improve command verification for Kenwood & Elecraft
Allows for a single ';' for  verification with the Elecraft XG3 single
character commands.

Allow for rig specific instance data  (needed by the Elecraft XG3 that
is a bit different from the rest of the Kenwood/Elecraft group).

Do not send a startup command to  turn off AI mode to the Elecraft XG3
as it doesn't have that command or mode.

Restore the positive RTTY mode enabled  check to the Elecraft K2 as it
appears that the MD6; and MD9;  comamnds are valid even when RTTY mode
is disabled on the rig.