Fixes to termios.c on W32 targets
I was debugging termios.c and this patch fixes two small problems--two
debug printf format problems and another one that causes an extra char
to be read on timeout.

The extra char problem is easily duplicated by using the "w" command
under rigtctl.  Most all other commands use fixed-length response
strings which don't exhibit the problem.  But since "w" does not know
how long the response should be it loops until it times out.  Every time
it times out it would return 1 extra character (this was 100% repeatable
on two different rigs).

Turns out it was a simple typo in the code as the read was using the
write event instead of the read event.