Allow for PowerSDR ID in Kenwood TS2000 emulation mode
PowerSDR  accepts  a  ZZID;  command   which  changes  the  result  of
subsequent ID; commands to ID900;. Since there seems no way of telling
PowerSDR to revert to  the true TS2000 ID of ID019;  we have to accept
ID900; as a TS2000.

Allow for  SmartSDR Kenwood  compatible IDs  of 904  - 907  as TS-2000
synonyms.