Command validation for Yaesu set commands
As with  Kenwood, the new  style Yaesu rigs  do not ACK  set commands.
Rather than  wait for  an error,  it is  quicker to  send a  follow up
simple query (e.g. ID;) and read the  reply, if it is a valid response
then the original set command worked; if it is a NACK response e.g. ?;
then  the original  command  failed. Use  this  mechanism to  initiate
retries of the original command.

(cherry picked from commit 1a02431da58131e65d54aba8d07846c12f5aa3b3)