Add preferred astylerc configuration
Astyle is a code formatting utility that can quickly reformat a source
file based on the settings of its configuration file.  The added
configuration file will format to a style mostly consistent with the
Linux Kernel Style guide:

https://www.kernel.org/doc/Documentation/CodingStyle

While conformance to this document is not the goal, it supplies a
reasonable set of guidelines to be used in Hamlib.