Do not swap VFO unnecessarily
Do the minimum swapping and other manipulation to determine the Tx and
VFO status when setting the split frequency and mode.

TODO - other funcitons.

The strategy  to determine the  current VFO at the  start of a  set or
query sequence of the split transmit VFO should not switch back to the
Rx VFO unless  it is certain that  it was current at the  start of the
sequence.