Use a sensible port timeout value for the Kenwood TS-850
This  back end  used a  480  second timeout  which effectively  blocks
retries.  The  long timeout  claimed to  be for  something to  do with
tuning  which I  assume is  related to  an auto  ATU blocking  rig CAT
responses.  The standard retry mechanism should cope with this without
resorting to an artificially extended timeout.