Icom frame patch
I'm implementing several changes to support some functionality for the
IC7300.  In order to minimize the review process I'm submitting this
small one separately (still working on finish the IC7300 but it's some
fairly extensive changes since ICOM has changed the protocol for several
things).

For the IC7300 turning power on requires ~150 0xfe's be sent to wake up
the rig.  This causes a buffer overflow condition in icom/frame.c.  This
patch increases the buffer size and ensures no overflow.

73
Mike W9MDB