Fix issue with TS-570S|D open in rig_caps
If the ts570 backend was called without the radio being turned on or the
wrong IO port specified, the library would generate a number of timeouts
and then fall through to behaving as though the communications channel
were opened.  Discovered that the backend did not specify a function for
rig_open in the rig_caps structure.

Thanks to Volker Schroer, DL1KSV, for reporting this issue.