Revised kenwood rig busy response handling
The TS-790  is very prone  to slow  command responses, this  change no
longer  retries  the sending  of  the  original  command when  a  busy
response "?;"  is seen, instead  it simply  waits one retry  delay and
tries response  read again. It  may be that  this has always  been the
correct action  but most rigs do  not return busy responses  or behave
better when a command is retried.