Simplify and fix defects in the FT991 back end
The FT991 only receives  on VFO A, VFO B is only used  for split Tx or
for exchanging with VFO A. There are no commands to set or query VFO B
mode. Added a set split mode function  that saves VFO A mode and VFO b
frequency then sets the required VFO B mode into VFO A and then copies
VFO  A  to  VFO  B  and  restores  the  VFO  B  frequency  and  VFO  A
mode. Bandwidth and narrow settings are not touched.

Several functins that had been  implemented specifically for the FT991
have been removed because existing newcat.c functions were suitable or
there should not have been a function as the rig doesn't support it.