Fix DIG MODE for FT817-ND
The FT817-ND only has RTTY, PSK31-L, PSK31-U, USER-L, USER-U as
possible DIG MODEs.  Before this fix, the modes read by Hamlib were
completely wrong.

The RTTY mode in the FT817-ND is USB. I think that the corresponding
Hamlib mode is RIG_MODE_RTTYR but I'm not certain if this is supposed to
be USB or LSB.

This has been tested on an FT817-ND. The FT817 might be different.