Allow RIG_LEVEL_STRENGTH queries over the network link
The rig  frontend may be  able to emulate RIG_LEVEL_STRENGTH  by using
RIG_LEVEL_RAWSTR and an S-meter  calibration curve.  The netrig client
doesn't know  this so it must  pass STRENGTH queries regardless  if it
knows a RIG_LEVEL_RAWSTR query is allowed. If the server cannot oblige
it will return a -RIG_EINVAL error which should be fine.