FT-817 status queries honour retries and timeout
This is  to help  with owners  of Arduino boards  like the  ubitx that
emulate an  FT-817 but respond badly  to a DTR signal  glitch which is
inevitable when opening a Linux  serial port.  The glitch triggers the
Arduino  boot-loader which  must be  allowed  to time  out before  the
FT-817 emulation starts.