Simplify SWIG interface definition and hide 64-bit values from Lua binding
By  including the  stdint  SWIG  typemaps we  can  directly use  types
derived from [u]int64_t etc.

Because Lua numbers are  implemented using a double-precision floating
point type and  also because SWIG when generating a  Lua wrappings can
only define  constants from #define and  enum values that fit  into an
'int' type we  must hide any constants that require  more than 32-bits
to represent themselves,  as they cannot be represented  in Lua.  This
applies  to  rmode_t  and  the  RIG_FUNC... defines  at  the  time  of
writing. By  hiding them from  the Lua binding  they will get  a 'nil'
value in Lua i.e. undefined so clients using the Lua binding will know
they are  not there.  This is  a nasty hack  but without  major design
changes to the Hamlib API it is do this or drop the Lua binding.