Backup serial options on open, restore on close
Before altering the serial port settings, make a copy of the settings
and track it by the fd. On serial port close, restore the original
settings for that port.

On Mac OS the Prolific PL2032 kernel driver locks up if closing without
first restoring the options. When this happens, further usage of the port
is impossible until computer is rebooted, which cannot even be done
without a hard power-off.
With this fix, a clean close make sure we can re-open the port again.