Remove GNU Texinfo dependency
As all of the information in the Texinfo manual had been converted to
Unix manual pages some time ago, remove GNU Texinfo as a distribution
dependency.  The rationale is that manual pages (roff) has more support
in various editors than Texinfo and the man utility is rather well
understood in contrast to the poor quality of the stand-alone GNU info
utility.  Several utilities are available to convert the roff man pages
into various formats such as HTML and PDF.