Fix spurious writes when no RT-21 second port
The if condition for whether there actually is a second port is only
covering the sprintf for the elevation command, and not the write
attempt. So we call rotorez_send_priv_cmd2 whether rotport2 is valid or
not. When it's not, we try to write the command to fd 0 (stdin). When
running on the console, this results in trash on the console. When
runnning under something else (e.g. systemd) this results in errors or
other wierd behavior (e.g. EBADF because stdin isn't connected, causing
the command to fail with RIG_EIO).