Fix sending morse with QRPLabs QCX rigs
The QCX series emulates the Kenwood TS480. However for the sending
morse command it seems to be more in line with the Elecraft K series
rigs.

Before this change, the QCX mini would reply to the KY command with
KY2 which was not understood. The QCX mini defines the KY responses to
be:

- KY0; if message sending is in progress and the transmit text buffer is not more than 75% full
- KY1; if message sending is in progress and the transmit text buffer is more than 75% full
- KY2; if no message is being sent, therefore the transmit text buffer is emtpy

This seems to be an extension of the original Kenwood protocol but it
is exactly as implemented by Elecraft in their K series rigs.

This patch treats the KY2 response the same as the KY0 response
meaning - OK to transmit.