Fix kendwood_priv_data definition
When compiling Hamlib 4.6 on Alpine Linux, the build fails with the
following error message:

    kenwood.h:186:5: error: unknown type name 'mode_t'; did you mean 'rmode_t'?

According to the man page for mode_t (on Debian), the standard C library
provides this type and it is an integer type. The struct member
last_mode_pc is compared with curr_mode which has type rmode_t so this
commit updates the definition to use this type.

With this change, Hamlib 4.6 builds successfully on Alpine Linux 3.21.0.